Feat : adding a work manager to enqueue sync process in background.
Propagate a callback every ack, in order to persist event, (ext. resp, not implemented), to avoid the in-app-memory and event to be lost. Returning chunks give us the possibility to implement a recovery strategy in case the app is force closed

Propagate an in-progress int to the UI

Making the code more kotlin idiomatic.